add option to keep log

+++ mgen-simulator.sh
@@ -169,6 +169,7 @@ if [[ $client ]]; then
 
 	analyze_latency_jitter_mgen_seq -v nflows=$NUM_FLOWS -v pps=$pack_per_second -v dur=$sec -v size=$bytes_per_packet $outfile
 
-	rm $outfile
-
+	if [[ ! $keep_drc ]]; then
+		rm $outfile
+	fi
 fi
